Some changes to timeseries for data hiding.  Added darvas scanning to bulk query.
git-svn-id: svn+ssh://svn.oneunified.net/var/local/svn/private/projects/oneunified/projects/trunk/TradeFrame@400 43143c16-083f-4785-a0a0-245a1fb241a9

+++ LibTimeSeries/TimeSeries.cpp
@@ -57,7 +57,7 @@ CDatedDatums::~CDatedDatums( void ) {
 CBars::CBars(void) {
 }
 
-CBars::CBars( unsigned int size )
+CBars::CBars( size_t size )
 :CTimeSeries<CBar>( size ) {
 }
